social_engineering_simulator/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
spacy>=3.7.0
